requirements.txt
django
djangorestframework
django-cors-headers
whitenoise
gunicorn
psycopg2-binary
inflect